NTBS-572: Move files to new folder structure
