+++ ntbs-service/Pages/Notifications/Edit/Items/_TreatmentEvent.cshtml
